requirements.txt
ultralytics==8.3.0
torch>=2.0.0
torchvision>=0.15.0
roboflow==1.1.23
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.31.0
opencv-python==4.9.0.80
pillow==10.2.0
numpy==1.26.3
pandas==2.2.0
pyyaml==6.0.1
python-multipart==0.0.6
aiofiles==23.2.1
tqdm==4.66.1
